Typo

+++ a2/thrusterSubsystem.c
@@ -17,7 +17,7 @@ void thrusterSubsystem(void *thrustStruct) {
   parseCommands(thrusterCommand, &cc);
 
   // 3. Calculate the fuel cost
-  unsigned double fuelCost = getFuelCost(&cc);
+  double fuelCost = getFuelCost(&cc);
 
   // 4. Update the fuelLvl
   *fuelLvl -= (unsigned short)fuelCost;
